+++ backend/src/main/java/org/ftclub/cabinet/auth/domain/OauthLink.java
@@ -1,4 +1,4 @@
-package org.ftclub.cabinet.oauth.domain;
+package org.ftclub.cabinet.auth.domain;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
