+++ backend/src/main/java/org/ftclub/cabinet/mapper/UserMapper.java
@@ -15,7 +15,7 @@ import org.ftclub.cabinet.dto.UserBlockedInfoDto;
 import org.ftclub.cabinet.dto.UserOauthConnectionDto;
 import org.ftclub.cabinet.dto.UserProfileDto;
 import org.ftclub.cabinet.dto.UserProfilePaginationDto;
-import org.ftclub.cabinet.oauth.domain.OauthLink;
+import org.ftclub.cabinet.auth.domain.OauthLink;
 import org.ftclub.cabinet.user.domain.BanHistory;
 import org.ftclub.cabinet.user.domain.LentExtension;
 import org.ftclub.cabinet.user.domain.User;
