+++ backend/src/main/java/org/ftclub/cabinet/openpublic/controller/LoginController.java
@@ -7,8 +7,8 @@ import lombok.RequiredArgsConstructor;
 import org.ftclub.cabinet.alarm.config.AlarmProperties;
 import org.ftclub.cabinet.auth.service.AuthFacadeService;
 import org.ftclub.cabinet.log.Logging;
-import org.ftclub.cabinet.oauth.domain.OauthResult;
-import org.ftclub.cabinet.oauth.service.OauthFacadeService;
+import org.ftclub.cabinet.auth.domain.OauthResult;
+import org.ftclub.cabinet.auth.service.OauthFacadeService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
