+++ backend/src/main/java/org/ftclub/cabinet/security/handler/CustomSuccessHandler.java
@@ -9,9 +9,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.ftclub.cabinet.auth.service.AuthFacadeService;
 import org.ftclub.cabinet.auth.service.AuthPolicyService;
-import org.ftclub.cabinet.oauth.domain.CustomOAuth2User;
-import org.ftclub.cabinet.oauth.domain.OauthResult;
-import org.ftclub.cabinet.oauth.service.OauthFacadeService;
+import org.ftclub.cabinet.auth.domain.CustomOAuth2User;
+import org.ftclub.cabinet.auth.domain.OauthResult;
+import org.ftclub.cabinet.auth.service.OauthFacadeService;
 import org.ftclub.cabinet.security.exception.SecurityExceptionHandlerManager;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
