+++ backend/src/main/java/org/ftclub/cabinet/utils/leave/absence/LeaveAbsenceManager.java
@@ -6,8 +6,8 @@ import lombok.extern.log4j.Log4j2;
 import org.ftclub.cabinet.auth.domain.FtRole;
 import org.ftclub.cabinet.auth.service.ApplicationTokenManager;
 import org.ftclub.cabinet.lent.service.LentFacadeService;
-import org.ftclub.cabinet.oauth.domain.FtOauthProfile;
-import org.ftclub.cabinet.oauth.service.OauthProfileService;
+import org.ftclub.cabinet.auth.domain.FtOauthProfile;
+import org.ftclub.cabinet.auth.service.OauthProfileService;
 import org.ftclub.cabinet.user.service.UserCommandService;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
